WIP - API integration
pulling data for the Portfolio page is broken right now, as is Dropdown.test.js.

Filter.test.js and Table.test.js are updated and API integration has started. Login and Register endpoints are working but the rest are giving a 403 error and need to be debugged.

Moved the check for whether the filter is empty to the portfolio helpers file so it can be used from the Portfoliio and Filter components